Move drawings modal into separate component
